Market GUI stuff
- Green glass on selection
- Show if there are goods to collect on My Orders button
- Stats are now integrated
- Added more stats
- Trades are now integrated

+++ ProtoBusiness/src/main/java/net/pixlies/business/market/orders/Trade.java
@@ -21,7 +21,7 @@ import java.util.UUID;
 )
 public class Trade {
 
-    @Id private int timestamp;
+    @Id private long timestamp;
     private double price;
     private int amount;
